moved Renderer.js to renderer/index.js; add bundleColumns
