+++ src/Hypergrid.js
@@ -12,7 +12,7 @@ var _ = require('object-iterators'); // fyi: installs the Array.prototype.find p
 var Base = require('./Base');
 var defaults = require('./defaults');
 var Canvas = require('./lib/Canvas');
-var Renderer = require('./lib/Renderer');
+var Renderer = require('./renderer');
 var SelectionModel = require('./lib/SelectionModel');
 var stylesheet = require('./lib/stylesheet');
 var Localization = require('./lib/Localization');
